Generate German dictionary for JetBrains IDEs
Steps:

    $ brew install aspell --with-lang-de
    $ mkdir ~/.dotfiles/dicts
    $ aspell -l de_DE dump master > ~/.dotfiles/dicts/de_DE.dic
    # Set custom dictionary directory for example in RubyMine
    # in Preferences > Editor > Spelling > Dictionaries Tab

JetBrains IDEs expect dictionaries to be a list of words separated by newlines and this is exactly what the aspell dump creates. The dump of the German dictionary has 312002 entries and a size of 4 MB. It already includes the expanded entries (e.g. singular and plural forms) and is UTF-8 encoded (at least when created in a shell with LANG=en_US.UTF-8 and LC_CTYPE=en_US.UTF-8.